Module mpd::error

This module defines different errors occurring during communication with MPD.

There're following kinds of possible errors:

IO errors (due to network communication failures),

parsing errors (because of bugs in parsing server response),

protocol errors (happen when we get unexpected data from server, mostly because protocol version mismatch, network data corruption or just bugs in the client),

server errors (run-time errors coming from MPD due to some MPD errors, like database failures or sound problems)

This module defines all necessary infrastructure to represent these kinds or errors.

Structs

ServerError
Server error

Enums

Error
Main error type, describing all possible error classes for the crate

ErrorCode
Server error codes, as defined in libmpdclient

ParseError
Parsing error kinds

ProtoError
Protocol errors

Type Definitions

Result
Shortcut type for MPD results