Enum mpd::idle::Subsystem

pub enum Subsystem {
    Database,
    Update,
    Playlist,
    Queue,
    Player,
    Mixer,
    Output,
    Options,
    Sticker,
    Subscription,
    Message,
}

Subsystems for idle command

Variants
Database

database: the song database has been modified after update.

Update

update: a database update has started or finished. If the database was modified during the update, the database event is also emitted.

Playlist

stored_playlist: a stored playlist has been modified, renamed, created or deleted

Queue

playlist: the current playlist has been modified

Player

player: the player has been started, stopped or seeked

Mixer

mixer: the volume has been changed

Output

output: an audio output has been enabled or disabled

Options

options: options like repeat, random, crossfade, replay gain

Sticker

sticker: the sticker database has been modified.

Subscription

subscription: a client has subscribed or unsubscribed to a channel

Message

message: a message was received on a channel this client is subscribed to; this event is only emitted when the queue is empty