Module mpd::message

The module defines structures for MPD client-to-client messaging/subscription protocol

The MPD client-to-client messaging protocol is fairly easy one, and is based on channels. Any client can subscribe to arbitrary number of channels, and some other client can send messages to a channel by name. Then, at some point of time, subscribed client can read all queued messages for all channels, it was subscribed to.

Also client can get asynchronous notifications about new messages from subscribed channels with idle command, by waiting for message subsystem events.

Structs

Channel
Channel

Message
Message