Module mpd::mount

The module describes data structures for MPD (virtual) mounts system

This mounts has nothing to do with system-wide Unix mounts, as they are implemented inside MPD only, so they doesn't require root access.

The MPD mounts are plugin-based, so MPD can mount any resource as a source of songs for its database (like network shares).

Possible, but inactive, mounts are named "neighbors" and can be listed with neighbors() method.

Structs

Mount
Mount point

Neighbor
Neighbor