Struct mpd::client::Client

pub struct Client<S = TcpStream> where
    S: Read + Write,  {
    pub version: Version,
    // some fields omitted
}

Client connection

Fields
version: Version

MPD version

Methods

impl Client<TcpStream>

pub fn connect<A: ToSocketAddrs>(addr: A) -> Result<Client<TcpStream>>

Connect client to some IP address

impl<S: Read + Write> Client<S>

pub fn new(socket: S) -> Result<Client<S>>

Create client from some arbitrary pre-connected socket

pub fn status(&mut self) -> Result<Status>

Get MPD status

pub fn stats(&mut self) -> Result<Stats>

Get MPD playing statistics

pub fn clearerror(&mut self) -> Result<()>

Clear error state

pub fn volume(&mut self, volume: i8) -> Result<()>

Set volume

pub fn repeat(&mut self, value: bool) -> Result<()>

Set repeat state

pub fn random(&mut self, value: bool) -> Result<()>

Set random state

pub fn single(&mut self, value: bool) -> Result<()>

Set single state

pub fn consume(&mut self, value: bool) -> Result<()>

Set consume state

pub fn crossfade<T: ToSeconds>(&mut self, value: T) -> Result<()>

Set crossfade time in seconds

pub fn mixrampdb(&mut self, value: f32) -> Result<()>

Set mixramp level in dB

pub fn mixrampdelay<T: ToSeconds>(&mut self, value: T) -> Result<()>

Set mixramp delay in seconds

pub fn replaygain(&mut self, gain: ReplayGain) -> Result<()>

Set replay gain mode

pub fn play(&mut self) -> Result<()>

Start playback

pub fn switch<T: ToQueuePlace>(&mut self, place: T) -> Result<()>

Start playback from given song in a queue

pub fn next(&mut self) -> Result<()>

Switch to a next song in queue

pub fn prev(&mut self) -> Result<()>

Switch to a previous song in queue

pub fn stop(&mut self) -> Result<()>

Stop playback

pub fn toggle_pause(&mut self) -> Result<()>

Toggle pause state

pub fn pause(&mut self, value: bool) -> Result<()>

Set pause state

pub fn seek<T: ToSeconds, P: ToQueuePlace>(
&mut self,
place: P,
pos: T
) -> Result<()>

Seek to a given place (in seconds) in a given song

pub fn rewind<T: ToSeconds>(&mut self, pos: T) -> Result<()>

Seek to a given place (in seconds) in the current song

pub fn songs<T: ToQueueRangeOrPlace>(&mut self, pos: T) -> Result<Vec<Song>>

List given song or range of songs in a play queue

pub fn queue(&mut self) -> Result<Vec<Song>>

List all songs in a play queue

pub fn currentsong(&mut self) -> Result<Option<Song>>

Get current playing song

pub fn clear(&mut self) -> Result<()>

Clear current queue

pub fn changes(&mut self, version: u32) -> Result<Vec<Song>>

List all changes in a queue since given version

pub fn push<P: ToSongPath>(&mut self, path: P) -> Result<Id>

Append a song into a queue

pub fn insert<P: ToSongPath>(&mut self, path: P, pos: usize) -> Result<usize>

Insert a song into a given position in a queue

pub fn delete<T: ToQueueRangeOrPlace>(&mut self, pos: T) -> Result<()>

Delete a song (at some position) or several songs (in a range) from a queue

pub fn shift<T: ToQueueRangeOrPlace>(
&mut self,
from: T,
to: usize
) -> Result<()>

Move a song (at a some position) or several songs (in a range) to other position in queue

pub fn swap<T: ToQueuePlace>(&mut self, one: T, two: T) -> Result<()>

Swap to songs in a queue

pub fn shuffle<T: ToQueueRange>(&mut self, range: T) -> Result<()>

Shuffle queue in a given range (use .. to shuffle full queue)

pub fn priority<T: ToQueueRangeOrPlace>(
&mut self,
pos: T,
prio: u8
) -> Result<()>

Set song priority in a queue

pub fn range<T: ToSongId, R: ToSongRange>(
&mut self,
song: T,
range: R
) -> Result<()>

Set song range (in seconds) to play

Doesn't work for currently playing song.

pub fn tag<T: ToSongId>(
&mut self,
song: T,
tag: &str,
value: &str
) -> Result<()>

Add tag to a song

pub fn untag<T: ToSongId>(&mut self, song: T, tag: &str) -> Result<()>

Delete tag from a song

pub fn ping(&mut self) -> Result<()>

Just pings MPD server, does nothing

pub fn close(&mut self) -> Result<()>

Close MPD connection

pub fn kill(&mut self) -> Result<()>

Kill MPD server

pub fn login(&mut self, password: &str) -> Result<()>

Login to MPD server with given password

pub fn playlists(&mut self) -> Result<Vec<Playlist>>

List all playlists

pub fn playlist<N: ToPlaylistName>(&mut self, name: N) -> Result<Vec<Song>>

List all songs in a playlist

pub fn load<T: ToQueueRange, N: ToPlaylistName>(
&mut self,
name: N,
range: T
) -> Result<()>

Load playlist into queue

You can give either full range (..) to load all songs in a playlist, or some partial range to load only part of playlist.

pub fn save<N: ToPlaylistName>(&mut self, name: N) -> Result<()>

Save current queue into playlist

If playlist with given name doesn't exist, create new one.

pub fn pl_rename<N: ToPlaylistName>(
&mut self,
name: N,
newname: &str
) -> Result<()>

Rename playlist

pub fn pl_clear<N: ToPlaylistName>(&mut self, name: N) -> Result<()>

Clear playlist

pub fn pl_remove<N: ToPlaylistName>(&mut self, name: N) -> Result<()>

Delete playlist

pub fn pl_push<N: ToPlaylistName, P: ToSongPath>(
&mut self,
name: N,
path: P
) -> Result<()>

Add new songs to a playlist

pub fn pl_delete<N: ToPlaylistName>(&mut self, name: N, pos: u32) -> Result<()>

Delete a song at a given position in a playlist

pub fn pl_shift<N: ToPlaylistName>(
&mut self,
name: N,
from: u32,
to: u32
) -> Result<()>

Move song in a playlist from one position into another

pub fn rescan(&mut self) -> Result<u32>

Run database rescan, i.e. remove non-existing files from DB as well as add new files to DB

pub fn update(&mut self) -> Result<u32>

Run database update, i.e. remove non-existing files from DB

pub fn find<W>(&mut self, query: &Query, window: W) -> Result<Vec<Song>> where
W: Into<Window>,

Find songs matching Query conditions.

pub fn search<W>(&mut self, query: &Query, window: W) -> Result<Vec<Song>> where
W: Into<Window>,

Case-insensitively search for songs matching Query conditions.

pub fn list(&mut self, term: &Term, query: &Query) -> Result<Vec<String>>

Lists unique tags values of the specified type for songs matching the given query.

pub fn findadd(&mut self, query: &Query) -> Result<()>

Find all songs in the db that match query and adds them to current playlist.

pub fn lsinfo<P: ToSongPath>(&mut self, path: P) -> Result<Song>

Lists the contents of a directory.

pub fn outputs(&mut self) -> Result<Vec<Output>>

List all outputs

pub fn output<T: ToOutputId>(&mut self, id: T, state: bool) -> Result<()>

Set given output enabled state

pub fn out_disable<T: ToOutputId>(&mut self, id: T) -> Result<()>

Disable given output

pub fn out_enable<T: ToOutputId>(&mut self, id: T) -> Result<()>

Enable given output

pub fn out_toggle<T: ToOutputId>(&mut self, id: T) -> Result<()>

Toggle given output

pub fn music_directory(&mut self) -> Result<String>

Get current music directory

pub fn commands(&mut self) -> Result<Vec<String>>

List all available commands

pub fn notcommands(&mut self) -> Result<Vec<String>>

List all forbidden commands

pub fn urlhandlers(&mut self) -> Result<Vec<String>>

List all available URL handlers

pub fn tagtypes(&mut self) -> Result<Vec<String>>

List all supported tag types

pub fn decoders(&mut self) -> Result<Vec<Plugin>>

List all available decoder plugins

pub fn channels(&mut self) -> Result<Vec<Channel>>

List all channels available for current connection

pub fn readmessages(&mut self) -> Result<Vec<Message>>

Read queued messages from subscribed channels

pub fn sendmessage(&mut self, channel: Channel, message: &str) -> Result<()>

Send a message to a channel

pub fn subscribe(&mut self, channel: Channel) -> Result<()>

Subscribe to a channel

pub fn unsubscribe(&mut self, channel: Channel) -> Result<()>

Unsubscribe to a channel

pub fn mounts(&mut self) -> Result<Vec<Mount>>

List all (virtual) mounts

These mounts exist inside MPD process only, thus they can work without root permissions.

pub fn neighbors(&mut self) -> Result<Vec<Neighbor>>

List all network neighbors, which can be potentially mounted

pub fn mount(&mut self, path: &str, uri: &str) -> Result<()>

Mount given neighbor to a mount point

The mount exists inside MPD process only, thus it can work without root permissions.

pub fn unmount(&mut self, path: &str) -> Result<()>

Unmount given active (virtual) mount

The mount exists inside MPD process only, thus it can work without root permissions.

pub fn sticker(&mut self, typ: &str, uri: &str, name: &str) -> Result<String>

Show sticker value for a given object, identified by type and uri

pub fn set_sticker(
&mut self,
typ: &str,
uri: &str,
name: &str,
value: &str
) -> Result<()>

Set sticker value for a given object, identified by type and uri

pub fn delete_sticker(&mut self, typ: &str, uri: &str, name: &str) -> Result<()>

Delete sticker from a given object, identified by type and uri

pub fn clear_stickers(&mut self, typ: &str, uri: &str) -> Result<()>

Remove all stickers from a given object, identified by type and uri

pub fn stickers(&mut self, typ: &str, uri: &str) -> Result<Vec<String>>

List all stickers from a given object, identified by type and uri

pub fn find_sticker(
&mut self,
typ: &str,
uri: &str,
name: &str
) -> Result<Vec<(String, String)>>

List all (file, sticker) pairs for sticker name and objects of given type from given directory (identified by uri)

pub fn find_sticker_eq(
&mut self,
typ: &str,
uri: &str,
name: &str,
value: &str
) -> Result<Vec<String>>

List all files of a given type under given directory (identified by uri) with a tag set to given value