Enum mpd::error::ParseError

pub enum ParseError {
    BadInteger(ParseIntError),
    BadFloat(ParseFloatError),
    BadValue(String),
    BadTime(TimeParseError),
    BadVersion,
    NotAck,
    BadPair,
    BadCode,
    BadPos,
    NoCodePos,
    NoMessage,
    NoRate,
    NoBits,
    NoChans,
    BadRate(ParseIntError),
    BadBits(ParseIntError),
    BadChans(ParseIntError),
    BadState(String),
    BadErrorCode(usize),
}

Parsing error kinds

Variants

BadInteger

invalid integer

BadFloat

invalid float

BadValue

some other invalid value

BadTime

date/time parsing error

BadVersion

invalid version format (should be x.y.z)

NotAck

the response is not an ACK (not an error) (this is not actually an error, just a marker to try to parse the response as some other type, like a pair)

BadPair

invalid pair

BadCode

invalid error code in ACK response

BadPos

invalid command position in ACK response

NoCodePos

missing command position and/or error code in ACK response

NoMessage

missing error message in ACK response

NoRate

missing bitrate in audio format field

NoBits

missing bits in audio format field

NoChans

missing channels in audio format field

BadRate

invalid bitrate in audio format field

BadBits

invalid bits in audio format field

BadChans

invalid channels in audio format field

BadState

unknown state in state status field

BadErrorCode

unknown error code in ACK response